Support Async for fiber-based concurrency
