+++ sdk/ruby/Gemfile
@@ -12,5 +12,8 @@ gem 'rspec', '~> 3.0'
 gem 'debug'
 
 group :test do
+  # Async to test Datastar::AsyncExecutor
+  gem 'async'
+  # Puma to host test server
   gem 'puma'
 end
